bank/requirements.txt
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
psycopg2==2.8.3
pytz==2019.2
